fix: Show total parked time in certain views
- Move VehicleModelMetaData to its own file
- Add TotalTimeParked as an attribute on the ParkedVehicle model
- Display TotalTimeParked in Details and EndParking views
- Removed number of days from display of TotalTimeParked
