+++ Uppgift12-Garage20/ViewModels/VehicleSummaryViewModel.cs
@@ -23,7 +23,7 @@ namespace Uppgift12_Garage20.ViewModels
         /// </summary>
         /// <param name="vehicle">The ParkedVehicle to view.</param>
         public VehicleSummaryViewModel(ParkedVehicle vehicle)
-            {
+        {
             ParkedVehicleId = vehicle.ParkedVehicleId;
             VehicleType = vehicle.VehicleType;
             RegistrationNumber = vehicle.RegistrationNumber;
