+++ Uppgift12-Garage20/Views/ParkedVehicles/Details.cshtml
@@ -64,6 +64,12 @@
         <dd class = "col-sm-10">
             @Html.DisplayFor(model => model.ArrivalTime)
         </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.TotalParkingTime)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.TotalParkingTime)
+        </dd>
     </dl>
 </div>
 <div>
