+++ Uppgift12-Garage20/Views/ParkedVehicles/EndParking.cshtml
@@ -53,6 +53,12 @@
         <dd class="col-sm-10">
             @Html.DisplayFor(model => model.ArrivalTime)
         </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.TotalParkingTime)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.TotalParkingTime)
+        </dd>
     </dl>
 
     <form asp-action="EndParking">
